requirements.txt
fastapi
uvicorn
faiss-cpu
pandas
python-dotenv
requests
pydantic
tqdm
streamlit
